% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.r
\name{download_ebooks}
\alias{download_ebooks}
\title{Downloads Free Text Ebooks From Springer Website}
\usage{
download_ebooks(mode = "all")
}
\description{
This package downloads free text ebooks from Springer based on a PDF received via WhatsApp on April 30th 2020.
       The Package uses a CSV data file containing all required information to download and name the files.
       Ebooks are available in two formats: (1) PDF (all ebooks), (2) EPUB (some ebooks).
       Formats can be selected by parameter "mode" ("all" -default-, "pdf", "epub")
       Ebooks are stored in the working directory.
}
\examples{
download_ebooks(), download_ebooks(mode = "pdf"), download_ebooks(mode = "epub")

}
